refactor: centralize app initialization
